Add files via upload
The code first calculates participants' visits on each AOI, preprocessed in the Pupil Cloud platform using AprilTag markers. The experiment entails 4 participants sitting at the table, hence the fixation coordinates are flipped 180 degrees for 2 participants in order to detect joint visual attention. Pairwise matrices are calculated to analyze overlapping visits with distance and time proximity thresholds. A distance matrix represents Euclidean distance between the mean visit positions of each pair of participants. A boolean matrix of time proximity is calculated based on the differences between the start and end times of the visits, given the threshold. An adjacency matrix is used to construct a graph of events, which describes if the visits are connected into the same joint visual attention event, taking into account both their spatial and temporal proximity.